Add support for setting exception and interrupt handlers
Much of this code is taken from the `cortex-m-rt` crate and the `stm32f103xx`
crate. `cortex-m-rt` could have been used if it had included the Debug execption
handler. In addition, the linker script would have needed to be replaced to
allow setting the `.flashconfig` section to allow booting with the Teensy
bootloader.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>