sim: Add COP support
The COP is more or less the watchdog of the mkl26. This adds a seperate struct
for the COP, even though the COP *technically* is part of the SIM. This was done
because their functionality is not connected in any way and there is no clock
gating that needs to be worried about.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>