adc: fix for mkl26
The mkl26 family only has ADC0.

The default voltage reference is connected to external pins. On the Teensy LC
the postitive pin is not easily connected to for a breadboard. The alternative
voltage reference is internally tied to VDD - VSS and so that was used for
testing.

The clock divisor and resolution were chanced from raw ints to enums to make one
more thing (well, two in this case) a compile time check.

Some pins were missing from the list and were added.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>