Update crates to be able to not place RefCells in Mutexes
This relies on a change in `bare-metal` that may or may not get merged upstream.

This also required updating to `cortex-m-rt` 0.5 which brought many changes
including, but not limited to, changes to how interrupt handlers are set,
changes to how the main function and default handlers are set, and a new
implementation of a way to run code before RAM is initialized.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>