Add tests for the absolute addresses of all peripheral registers
The tests create a reg struct at the starting address for each peripheral. The
address of each field in the structs are then tested against the absolute
address of the corresponding register as provided by the reference manual.

These tests caught a couple errors in the placement of a handful of registers
that are not currently being used in the provided drivers. To cut down on more
possible errors, the address of each peripheral was added as a const.

The tests do not need to run on a microcontroller as no reads or writes to the
addresses of the registers are performed. This means that the tests can run on a
host machine if the right target triple is used. The only problem is that
currently when building the tests all of the examples are built even when the
cargo is told to skip them. The examples fail to build as they are missing
`eh_personality` functions which are required for unwinding which is required
for testing in the default test runner. The tests in the actual library do build
and can be run manually, but this does add an additional step to the process and
the name of the produced test binary is not static, possibly leading to the user
running the wrong test binary.

Some formatting was also performed as I was poking around.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>