tpm: Check if CPWMS is set on the module when setting channel modes
The reference manual doesn't say that non-PWM modes are valid with CPWMS set,
other than Software Compare, but this doesn't make tons of sense since Software
Compare doesn't care about CPWMS being set. The don't-care on Software Compare
may be an error as Software Compare is a subset of Output Compare, and that
specifies that CPWMS must be 0.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>