tpm: carry lifetime of parent TPM in Channel and move checking of already enabled channels to Channel::new
The move was required so that a Channel could not outlive a parent TPM. If this
had happened, writes to the channel registers could have happened while the TPM
was disabled resulting in a crash.

The mutability requrement of the `Tpm::channel` function was changed to allow
multiple channels to be made while tracking the lifetime of the parent Tpm.
It might be neccessary to change the mutability of `Tpm::set_period` to allow
the MOD register to be changed while channels are initialized.

The channel specific registers were also moved wholy to the ChannelRegs struct.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>