adc: Move channel control to a separate struct to allow channel changing
As the ADC only needs to be calibrated once after reset and then any channel can
be used, it makes no sense to fully bring down the ADC just to use another
channel.

Moving channel specific functions to a separate struct that mutably borrow the
ADC allows the channel to be changed while still keeping mutual exclusion. This
also removes the need for a completely separate `AdcDiff` struct that duplicated
a lot of code.

Signed-off-by: Gabriel Smith <ga29smith@gmail.com>